Fix training logic

+++ deepq.py
@@ -17,6 +17,7 @@ from torch.utils.tensorboard import SummaryWriter
 
 from gymenv import MancalaEnv
 from agent import Agent
+from simple import MaxAgent
 
 model_fn = os.path.join("save", "policy")
 REPORTING_PERIOD = 100
@@ -26,33 +27,35 @@ if torch.cuda.is_available():
     # GPU Config
 
     device = torch.device('cuda')
+    map_location = 'cuda'
 
     # Training settings
 
-    BATCH_SIZE = 32
-    GAMMA = 0.98
-    EPS_START = 1
-    EPS_END = 0.05
-    EPS_DECAY = 0.0000002
-    MEMORY_SIZE = 5000000
-    LR = 0.0005
-    UPDATE_TARGET = 1500
-
-else:
-    # CPU Config
-
-    device = torch.device('cpu')
-
-    # Training settings
-
-    BATCH_SIZE = 16
+    BATCH_SIZE = 8
     GAMMA = 0.98
     EPS_START = 1
     EPS_END = 0.01
     EPS_DECAY = 0.0000025
     MEMORY_SIZE = 2000000
-    LR = 0.000001
-    UPDATE_TARGET = 2500
+    LR = 0.0001
+    UPDATE_TARGET = 4000
+
+else:
+    # CPU Config
+
+    device = torch.device('cpu')
+    map_location = 'cpu'
+
+    # Training settings
+
+    BATCH_SIZE = 8
+    GAMMA = 0.98
+    EPS_START = 1
+    EPS_END = 0.01
+    EPS_DECAY = 0.0000025
+    MEMORY_SIZE = 2000000
+    LR = 0.0001
+    UPDATE_TARGET = 5000
 
 ZEROED_ACTION_MASK = torch.zeros((BATCH_SIZE, 6)).to(device)
 
@@ -222,11 +225,12 @@ if __name__ == '__main__':
 
     if model_fn is not None and os.path.isfile(model_fn):
         print("Resuming from checkpoint: {} ...".format(model_fn))
-        policy_net = torch.load(os.path.join(os.getcwd(), model_fn), map_location='cpu')
+        policy_net = torch.load(os.path.join(os.getcwd(), model_fn), map_location=map_location)
     else:
         policy_net = MancalaAgentModel().to(device)
 
     agent = DeepQAgent(strategy, device, policy_net)
+    maxagent = MaxAgent()
     policy_net = agent.policy_net
 
     memory = ReplayMem(MEMORY_SIZE)
@@ -248,49 +252,51 @@ if __name__ == '__main__':
 
     while 1:
 
-        state = env.reset()
-
         n_episodes += 1
-        ep_reward = 0
+        ep_reward_model = 0
+
+        state = env.reset()
+        player_1_last_state = None
+        player_2_last_state = None
 
         for timestep in count():
 
             valid_actions = env.get_valid_actions()
 
             if env.active_player == 0:
-                action = agent.select_action(state, valid_actions, training_mode=True)
-                next_state, reward, done, info = env.step(action)
-                if not done:
-                    memory.push(Experience(state, action, next_state, reward))
-                else:
-                    # Return zero state when episode ends
-                    memory.push(Experience(state, action, np.zeros(14), reward))
+
+                if player_1_last_state is not None:
+                    memory.push(Experience(player_1_last_state, player_1_action, state, player_1_reward))
+
+                player_1_action = agent.select_action(state, valid_actions, training_mode=False)
+                next_state, player_1_reward, done, info = env.step(player_1_action)
+
+                player_1_last_state = state.copy()
+
+                ep_reward_model += player_1_reward
 
             else:
-                # Choose an action from player 2's perspective
-                action = agent.select_action(MancalaEnv.shift_view_p2(state), valid_actions, training_mode=True)
-                next_state, reward, done, info = env.step(action)
-                if not done:
-                    memory.push(Experience(
-                        MancalaEnv.shift_view_p2(state),
-                        action,
-                        MancalaEnv.shift_view_p2(next_state),
-                        reward
-                    )
-                )
-                else:
-                    # Return zero state when episode ends
-                    memory.push(Experience(
-                        MancalaEnv.shift_view_p2(state),
-                        action,
-                        np.zeros(14),
-                        reward
-                    )
-                )
 
-            ep_reward += reward
+                if player_2_last_state is not None:
+
+                    memory.push(Experience(
+                        MancalaEnv.shift_view_p2(player_2_last_state),
+                        player_2_action,
+                        MancalaEnv.shift_view_p2(state),
+                        player2_reward
+                        )
+                    )
+
+                # Choose an action from player 2's perspective
+
+                player_2_action = maxagent.select_action(MancalaEnv.shift_view_p2(state), valid_actions, env=env)
+                next_state, player2_reward, done, info = env.step(player_2_action)
+
+                player_2_last_state = state
 
             state = next_state
+
+            # Training
 
             if memory.can_provide_sample(BATCH_SIZE):
                 experiences = memory.sample(BATCH_SIZE)
@@ -310,9 +316,9 @@ if __name__ == '__main__':
                 n_batches_total += 1
                 n_batches_this_period += 1
 
-            if done:
+            # Handle end of episode
 
-                if n_episodes % REPORTING_PERIOD == 0:
+                if done and n_episodes % REPORTING_PERIOD == 0:
 
                     print("Episode {} completed, last {} episodes avg. duration: {}, avg. reward: {}".format(
                         n_episodes,
@@ -335,27 +341,32 @@ if __name__ == '__main__':
                     )
                     # Tensorboard reporting
 
-                    writer.add_scalar("Loss", total_loss / n_batches_this_period, n_episodes)
+                    writer.add_scalar("Training loss", total_loss / n_batches_this_period, n_episodes)
                     writer.add_scalar("Exploration rate", agent.get_exploration_rate(), n_episodes)
                     writer.add_scalar("Episode duration", np.mean(np.mean(episode_durations[-REPORTING_PERIOD:])),
                                       n_episodes)
+                    writer.add_scalar("Reward earned by model", ep_reward_model / REPORTING_PERIOD, n_episodes)
 
                     total_loss = 0
                     n_batches_this_period = 0
 
-                if n_episodes % UPDATE_TARGET == 0:
-                    print("Updating target net & saving checkpoint...")
-                    target_net.load_state_dict(policy_net.state_dict())
-
-                    if os.path.isfile(model_fn):
-                        os.remove(model_fn)
-                    torch.save(policy_net, model_fn)
-
                 writer.flush()
 
-                episode_durations.append(timestep)
-                episode_rewards.append(ep_reward)
+                if done:
 
-                break
+                    if n_episodes % UPDATE_TARGET == 0:
+                        print("Updating target net & saving checkpoint...")
+                        target_net.load_state_dict(policy_net.state_dict())
+
+                        if os.path.isfile(model_fn):
+                            os.remove(model_fn)
+                        torch.save(policy_net, model_fn)
+
+                        optimizer = optim.Adam(params=policy_net.parameters(), lr=LR)
+
+                    episode_durations.append(timestep)
+                    episode_rewards.append(ep_reward_model)
+
+                    break
